pyproject.toml
[tool.poetry]
name = "simple-async-command-manager"
version = "0.1.0"
description = "Simple command manager and asyncio command queue handler with wraps for subprocesses and synchronous commands. Designed to be modular for easy profiling and monitoring."
authors = ["SBroaddus220 <steven.broaddus220@gmail.com>"]
license = "GPL-3.0-only"
readme = "README.md"
packages = [{include = "simple_async_command_manager"}]

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = { version = "^20.4", optional = true }

[tool.poetry.dev-dependencies]
pytest = "^6.0"

[tool.poetry.extras]
telegram = ["python-telegram-bot"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
